Update to rabbitmq client v7 (#1830)
* upgrade packages and configure NuGetAudit

* update to net 8.0

* fix library build

need to update support for publish confirm and confirm select

+++ Source/EasyNetQ/ConnectionConfiguration.cs
@@ -148,7 +148,7 @@ public class ConnectionConfiguration
     /// If it is not set, a value of <seealso cref="PrefetchCount"/> is used due to backward compatibility
     /// </summary>
     /// <remarks>For concurrency greater than one, the consumers could process messages in any order, not in the order they receive them</remarks>
-    public int? ConsumerDispatcherConcurrency { get; set; } = null;
+    public ushort? ConsumerDispatcherConcurrency { get; set; } = null;
 }
 
 /// <summary>
